requirements.txt
# --- Core Agent Framework (Strict Pins) ---
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langchain-community==0.3.21
langchain-ollama==0.2.2
langgraph==0.3.4
# Explicitly pin the main package to match core 0.3.51
langchain==0.3.23

# --- LLM & Embeddings ---
# Note: Installed separately in script to handle dependency overlap
langchain-huggingface>=1.0.0
sentence-transformers>=3.0.0

# --- Vector Store ---
faiss-cpu>=1.8.0

# --- F1 Data & Analysis ---
fastf1>=3.4.0
pandas>=2.2.0
numpy>=2.0.0

# --- Visualization ---
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.20.0

# --- Web & API ---
requests>=2.32.0
python-dotenv>=1.0.0
backoff>=2.2.0
wikipedia>=1.4.0

# --- Async HTTP (NEW - Phase 1 Optimization) ---
aiohttp>=3.9.0           # Async HTTP client for parallel requests
httpx>=0.27.0            # Modern async HTTP with HTTP/2 support

# --- Performance Optimization (NEW) ---
orjson>=3.9.0            # Faster JSON parsing (2-3x speed)
diskcache>=5.6.0         # Persistent disk cache
memory-profiler>=0.61.0  # Memory usage profiling

# --- Document Processing ---
pypdf>=5.0.0
beautifulsoup4>=4.12.0

# --- CLI, Logging & Utils ---
rich>=13.0.0
termcolor>=2.4.0
tqdm>=4.67.0
python-dateutil>=2.9.0
loguru>=0.7.0
ffmpeg-python>=0.2.0
arcade>=3.0.0